Added some authorization to user follows & tweet creation.

+++ controllers/follow.go
@@ -4,6 +4,7 @@ import (
 	"database/sql"
 	"github.com/gorilla/mux"
 	_ "github.com/lib/pq"
+	"gowebapp/models"
 	"log"
 	"net/http"
 	"time"
@@ -29,14 +30,18 @@ func (ctrl *FollowController) Create(res http.ResponseWriter, req *http.Request)
 		log.Println(err)
 	}
 
-	_, err = ctrl.Db.Exec("insert into user_follows (follower_id, following_id, created_at) values ($1, $2, $3)", follower_id.Value, following_id, current_time)
+	if follower_id.Value != following_id {
+		_, err = ctrl.Db.Exec("insert into user_follows (follower_id, following_id, created_at) values ($1, $2, $3)", follower_id.Value, following_id, current_time)
 
-	if err != nil {
-		log.Println(err)
+		if err != nil {
+			log.Println(err)
+		}
 	}
+
 }
 
 func (ctrl *FollowController) Delete(res http.ResponseWriter, req *http.Request) {
+	follow := Models.Follow{}
 	url_params := mux.Vars(req)
 	following_id := url_params["user_id"]
 
@@ -46,9 +51,18 @@ func (ctrl *FollowController) Delete(res http.ResponseWriter, req *http.Request)
 		log.Println(err)
 	}
 
-	_, err = ctrl.Db.Exec("delete from user_follows where follower_id = $1 AND following_id = $2", follower_id.Value, following_id)
+	err = ctrl.Db.QueryRow("select id, follower_id, following_id, created_at from user_follows where follower_id = $1 and following_id = $2", follower_id.Value, following_id).Scan(&follow.Id, &follow.Follower_id, &follow.Following_id, &follow.Created_at)
 
-	if err != nil {
+	if err == sql.ErrNoRows {
+		log.Printf("\nno follow record found:  ")
 		log.Println(err)
+	} else if err != nil {
+		log.Println(err)
+	} else {
+		_, err = ctrl.Db.Exec("delete from user_follows where follower_id = $1 AND following_id = $2", follower_id.Value, following_id)
+
+		if err != nil {
+			log.Println(err)
+		}
 	}
 }
